+++ src/components/map/KakaoMap.tsx
@@ -1,19 +1,21 @@
 import { useEffect, useState } from 'react';
-import { CustomOverlayMap, Map, MapMarker } from 'react-kakao-maps-sdk';
+import { Map, MapMarker } from 'react-kakao-maps-sdk';
+import { useRecoilState } from 'recoil';
 import { mapMarkerDataHandler } from '../../api/Supabase.api';
+import { MarkerKakaoState } from '../../state/atom/kakaoMapAtom';
 import * as St from './KakaoMap.styled';
 import { LocationTypes, MapMarkerDataTypes, MarkerTypes, UserLocationStateTypes } from './KakaoMap.type';
 import AboutIndBookStore from './aboutindbookstore/AboutIndBookStore';
 import DetailMapInfo from './detailmapinfo/DetailInfo';
+import MapSearchInput from './mapsearchInput/MapSearchInput';
+import SelectMarkerInfo from './selectmarkerinfo/SelectMarkerInfo';
 
 function KakaoMap() {
   //db에 저장한 맵 마커 데이터
   const [markerData, setMarkerData] = useState<MapMarkerDataTypes[] | undefined>([]);
   const [location, setLocation] = useState<LocationTypes>({ latitude: null, longitude: null });
   const [error, setError] = useState<string | null>(null);
-  const [selectedMarkerId, setSelectedMarkerId] = useState<number | null>(null);
-  const [searchPlace, setSearchPlace] = useState<string>('');
-  const [markers, setMarkers] = useState<MarkerTypes[]>([]);
+  const [markers, setMarkers] = useRecoilState<MarkerTypes[]>(MarkerKakaoState);
   const [aroundStore, setAroundStore] = useState<MapMarkerDataTypes[]>([]);
   const [map, setMap] = useState<kakao.maps.Map | null>(null);
   const [selectedMarkerInfo, setSelectedMarkerInfo] = useState<MapMarkerDataTypes | null>(null);
@@ -21,9 +23,6 @@ function KakaoMap() {
     lat: 37.50231497199725,
     lng: 127.04484141806945,
   });
-  const baseUrl = process.env.REACT_APP_KAKAO_API_BASE_URL;
-
-  console.log(selectedMarkerInfo);
 
   useEffect(() => {
     if (navigator.geolocation) {
@@ -68,7 +67,6 @@ function KakaoMap() {
   useEffect(() => {
     if (map) {
       const zoomChangedCallback = () => {
-        const level = map.getLevel();
         fetchBookstoresNearby(map.getCenter().getLat(), map.getCenter().getLng());
       };
       kakao.maps.event.addListener(map, 'zoom_changed', zoomChangedCallback);
@@ -77,36 +75,6 @@ function KakaoMap() {
       };
     }
   }, [map]);
-
-  const searchStore = (e: React.FormEvent<HTMLFormElement>) => {
-    e.preventDefault();
-    const ps = new window.kakao.maps.services.Places();
-    if (searchPlace) {
-      ps.keywordSearch(searchPlace, (data, status, _pagination) => {
-        if (status === window.kakao.maps.services.Status.OK) {
-          const bounds = new kakao.maps.LatLngBounds();
-          let markers: MarkerTypes[] = data.map((item) => ({
-            position: {
-              lat: parseFloat(item.y),
-              lng: parseFloat(item.x),
-            },
-            content: item.place_name,
-          }));
-          setMarkers(markers);
-          if (data.length > 0) {
-            const firstResult = data[0];
-            const latitude = parseFloat(firstResult.y);
-            const longitude = parseFloat(firstResult.x);
-            fetchBookstoresNearby(latitude, longitude);
-          }
-          if (map) {
-            map.setBounds(bounds); // 지도의 확대/축소 레벨과 중심 위치를 설정합니다.
-          }
-        }
-      });
-    }
-    setSearchPlace('');
-  };
 
   // Express API에서 서점 목록을 가져오는 함수
   const fetchBookstoresNearby = async (latitude: number, longitude: number) => {
@@ -156,10 +124,6 @@ function KakaoMap() {
     }
   };
 
-  const onSearchChanged = (e: React.ChangeEvent<HTMLInputElement>) => {
-    setSearchPlace(e.target.value);
-  };
-
   const fetchMapMarkerData = async () => {
     try {
       const result = await mapMarkerDataHandler();
@@ -167,15 +131,6 @@ function KakaoMap() {
     } catch (err) {
       console.log(err);
     }
-  };
-
-  const markerClickHandler = (markerId: number) => {
-    setSelectedMarkerId((prev) => (prev === markerId ? null : markerId));
-  };
-
-  const selectMarkerInfoHandler = (markerId: number) => {
-    const clickedMarkerInfo = markerData?.find((marker) => marker.id === markerId);
-    setSelectedMarkerInfo(clickedMarkerInfo || null);
   };
 
   const updateMapCenter = (latitude: number, longitude: number) => {
@@ -191,28 +146,12 @@ function KakaoMap() {
         <St.Map>
           <St.Title>전국에 독립서점은 얼마나 있을까?</St.Title>
           {/* 검색창 */}
-          <St.SearchWrapper>
-            <St.SearchFormWrapper>
-              <St.SearchForm onSubmit={searchStore}>
-                <St.SearchBox>
-                  <St.SearchInput type="text" value={searchPlace} onChange={onSearchChanged} />
-                  <St.SearchButton>검색</St.SearchButton>
-                </St.SearchBox>
-              </St.SearchForm>
-            </St.SearchFormWrapper>
-            <St.SearchList>
-              {aroundStore &&
-                aroundStore.length > 0 &&
-                aroundStore.map((marker, index) => (
-                  <St.SearchResultWrapper
-                    key={index}
-                    onClick={() => updateMapCenter(marker.latitude, marker.longitude)}>
-                    <St.SearchName>{marker.name}</St.SearchName>
-                    <St.SearchAdress>{marker.address}</St.SearchAdress>
-                  </St.SearchResultWrapper>
-                ))}
-            </St.SearchList>
-          </St.SearchWrapper>
+          <MapSearchInput
+            fetchBookstoresNearby={fetchBookstoresNearby}
+            map={map}
+            aroundStore={aroundStore}
+            updateMapCenter={updateMapCenter}
+          />
           <Map
             id={'map'}
             center={currentPosition}
@@ -229,53 +168,7 @@ function KakaoMap() {
                 zIndex: 1,
               }}></MapMarker>
             {markerData?.map((position, index) => {
-              return (
-                <MapMarker
-                  key={`${position.name}-${position.latitude}`}
-                  position={{ lat: position.latitude, lng: position.longitude }}
-                  image={{
-                    src: 'https://t1.daumcdn.net/localimg/localimages/07/mapapidoc/markerStar.png', // 마커이미지의 주소입니다
-                    size: {
-                      width: 24,
-                      height: 35,
-                    }, // 마커이미지의 크기입니다
-                    options: {
-                      offset: {
-                        x: 27,
-                        y: 69,
-                      }, // 마커이미지의 옵션입니다. 마커의 좌표와 일치시킬 이미지 안에서의 좌표를 설정합니다.
-                    },
-                  }}
-                  onClick={() => markerClickHandler(position.id)}>
-                  {selectedMarkerId === position.id && ( //내가 연것만 열려야함
-                    <CustomOverlayMap
-                      zIndex={1}
-                      position={{ lat: position.latitude, lng: position.longitude }}
-                      yAnchor={1}>
-                      <St.InfoWrapper>
-                        <St.Info>
-                          <St.InfoTitle>
-                            {position.name}
-                            <div className="close" onClick={() => markerClickHandler(position.id)} title="닫기">
-                              x
-                            </div>
-                          </St.InfoTitle>
-                          <St.BodyWrapper>
-                            <St.ImgWrapper>
-                              <St.MapIndBookStoreImage />
-                            </St.ImgWrapper>
-                            <St.StoreInfoWrapper>
-                              <St.StoreInfo>{position.address}</St.StoreInfo>
-                              <St.StoreInfo>{position.tel_num}</St.StoreInfo>
-                            </St.StoreInfoWrapper>
-                          </St.BodyWrapper>
-                        </St.Info>
-                        <St.DetailButton onClick={() => selectMarkerInfoHandler(position.id)}>상세보기</St.DetailButton>
-                      </St.InfoWrapper>
-                    </CustomOverlayMap>
-                  )}
-                </MapMarker>
-              );
+              return <SelectMarkerInfo position={position} markerData={markerData} />;
             })}
           </Map>
         </St.Map>
@@ -283,8 +176,7 @@ function KakaoMap() {
           <></>
         ) : (
           <>
-            {' '}
-            <DetailMapInfo markerInfo={selectedMarkerInfo} />{' '}
+            <DetailMapInfo markerInfo={selectedMarkerInfo} />
           </>
         )}
       </St.MapContainer>
