+++ src/components/map/KakaoMap.type.ts
@@ -40,3 +40,37 @@ export type MapMarkerDataTypes = {
   adit_dc: string;
   postal_code: number;
 };
+
+export type KakaoProps = {
+  position: {
+    address: string;
+    adit_dc: string | null;
+    category1: string;
+    category2: string;
+    closedDay_close_at?: string | number;
+    closedDay_guide?: string | number;
+    closedDay_open_at?: string | number;
+    gov_id: string;
+    id: number;
+    latitude: number;
+    longitude: number;
+    name: string;
+    optn_dc: string | null;
+    postal_code: number | null;
+    saturday_close_at?: string | number;
+    saturday_open_at?: string | number;
+    sunday_close_at?: string | number;
+    sunday_open_at?: string | number;
+    tel_number?: number | null;
+    weekdays_close_at?: string | number;
+    weekdays_open_at?: string | number;
+  };
+  markerData: MapMarkerDataTypes[];
+};
+
+export type searchProps = {
+  fetchBookstoresNearby: (latitude: number, longitude: number) => void;
+  map: kakao.maps.Map | null;
+  aroundStore: MapMarkerDataTypes[];
+  updateMapCenter: (latitude: number, longitude: number) => void;
+};
